refactor: Update component imports for navbar in layout pages, change login route to signin
