+++ src/app/(auth)/signin/page.tsx
@@ -1,3 +1,4 @@
+import OauthSignin from '@/components/auth/oauth-signin'
 import LoginForm from '@/components/form/login.form'
 import Link from 'next/link'
 import React from 'react'
@@ -9,13 +10,14 @@ const LoginPage = () => {
                 <h2 className='text-3xl lg:text-5xl text-white font-bold font-poppins tracking-wide'>
                     Login your account.
                 </h2>
-                <p className='text-gray-300 text-sm lg:text-lg mt-3'>
-                    New here? <Link href="/register" className='hover:underline text-purple-400'>Create an account</Link>
+                <p className='text-gray-300 text-sm lg:text-lg mt-4'>
+                    New here? <Link href="/register" className='hover:underline text-purple-400'>Create an account</Link> or <Link className='hover:underline text-purple-400' href="/forgot-password">recover your lost account</Link>
                 </p>
             </div>
 
             <div className='mt-10'>
                 <LoginForm />
+                <OauthSignin />
             </div>
         </div>
     )
