+++ src/auth.ts
@@ -8,6 +8,11 @@ import { TAuthUser } from "./types/common";
 
 export const { handlers, signIn, signOut, auth } = NextAuth({
   providers: [
+    Google,
+    Linkedin({
+      clientId: process.env.LINKEDIN_ID,
+      clientSecret: process.env.LINKEDIN_SECRET,
+    }),
     Credentials({
       credentials: {
         email: { label: "Email", type: "email" },
@@ -45,14 +50,6 @@ export const { handlers, signIn, signOut, auth } = NextAuth({
         };
       },
     }),
-    Google({
-      clientId: process.env.GOOGLE_ID,
-      clientSecret: process.env.GOOGLE_SECRET,
-    }),
-    Linkedin({
-      clientId: process.env.LINKEDIN_ID,
-      clientSecret: process.env.LINKEDIN_SECRET,
-    }),
   ],
   callbacks: {
     authorized({ request: { nextUrl }, auth }) {
@@ -87,7 +84,7 @@ export const { handlers, signIn, signOut, auth } = NextAuth({
     },
   },
   pages: {
-    signIn: "/login",
+    signIn: "/signin",
     newUser: "/register",
   },
 });
